Update to MAPL v1.0.0 for several directories
Directories included in this update:
Chem_Shared, GEOS-Shared, GMAO_etc, GMAO_gfio, GMAO_hermes,
GMAO_mpeu, GMAO_pilgrim, and GMAO_perlib.

+++ Shared/GMAO_gfio/diffdate.f
@@ -29,7 +29,8 @@
 ! 
 ! !USES:
 !
-
+       use, intrinsic :: iso_fortran_env, only: INT64
+       
        implicit none
 
 !
@@ -75,7 +76,7 @@
        integer year1,mon1,day1,hour1,min1,sec1
        integer year2,mon2,day2,hour2,min2,sec2
 !      integer julian1, julian2, julsec1, julsec2
-       integer(kind=8) julian1, julian2, julsec1, julsec2
+       integer(kind=INT64) julian1, julian2, julsec1, julsec2
 
        character*8 dateString
